requirements.txt
arrow==1.3.0
asgiref==3.9.0
binaryornot==0.4.4
certifi==2025.7.9
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cookiecutter==2.6.0
diff-match-patch==20241021
Django==5.2.4
django-extensions==4.1
django-import-export==4.3.8
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==25.0
pillow==11.3.0
Pygments==2.19.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20250708
urllib3==2.5.0
psycopg[binary]
